requirements.txt
numpy
pandas
matplotlib
plotly
python-dotenv
openpyxl
sqlalchemy
